Move the volume guard to its own `DiagnosticLoggingVolumeTest` class (wala/ML#702).
The guard is an infrastructure test, not a tensor-model assertion, so it
doesn't belong in `TestTensorflow2Model`. The nlpgnn generation run it
reuses is extracted into a package-visible `runNlpgnnFullGeneration()`
helper, so the guard no longer invokes another class's `@Test` method.
