+++ com.ibm.wala.cast.python.ml.test/source/com/ibm/wala/cast/python/ml/test/TestTensorflow2Model.java
@@ -2183,6 +2183,21 @@ public class TestTensorflow2Model extends TestPythonMLCallGraphShape {
   @Test
   public void testNlpgnnFullGeneration()
       throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
+    runNlpgnnFullGeneration();
+  }
+
+  /**
+   * Runs the NLPGNN whole-project generation analysis with its call-graph and type assertions.
+   * Package-visible so {@link DiagnosticLoggingVolumeTest} can rerun the analysis under {@code
+   * FINEST} without invoking another class's {@code @Test} method.
+   *
+   * @throws ClassHierarchyException On WALA class-hierarchy error.
+   * @throws IllegalArgumentException On illegal argument.
+   * @throws CancelException On analysis cancellation.
+   * @throws IOException On I/O error reading the test file.
+   */
+  void runNlpgnnFullGeneration()
+      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
     test(
         NLPGNN_FULL_PROJECT_FILES,
         "tests/TG/EN/generation.py",
@@ -2215,62 +2230,6 @@ public class TestTensorflow2Model extends TestPythonMLCallGraphShape {
         1,
         1,
         Map.of(3, Set.of(new TensorType(UNKNOWN, asList(DynamicDim.INSTANCE, new NumericDim(1))))));
-  }
-
-  /**
-   * Diagnostic-logging volume guard (<a
-   * href="https://github.com/wala/ML/issues/702">wala/ML#702</a>). Reruns the nlpgnn generation
-   * analysis — the cyclic-closure call graph that triggered <a
-   * href="https://github.com/wala/ML/issues/697">wala/ML#697</a> — with every {@code
-   * com.ibm.wala.cast.python} logger at {@code FINEST}, routing records through {@link
-   * DiscardingFormattingHandler}, which formats and discards them while summing their volume.
-   *
-   * <p>Correct code renders diagnostics through the bounded {@code Loggables.describe(...)} (~0.6
-   * GB of formatted {@code FINEST} output for this analysis); a bare render of a {@code
-   * Context}-bearing value inflates that by more than an order of magnitude (~14 GB measured),
-   * which this bound catches. The failure is invisible at CI's {@code WARNING} level (the message
-   * strings are never built), so this test is the pipeline's guard against its return. See {@code
-   * CONTRIBUTING.md}'s Diagnostic Logging section and <a
-   * href="https://github.com/wala/WALA/issues/1992">wala/WALA#1992</a> for the upstream root cause.
-   *
-   * @throws ClassHierarchyException On WALA class-hierarchy error.
-   * @throws IllegalArgumentException On illegal argument.
-   * @throws CancelException On analysis cancellation.
-   * @throws IOException On I/O error reading the test file.
-   */
-  @Test
-  public void testDiagnosticLoggingVolumeBounded()
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    // Fixed code emits ~0.6 GB of formatted FINEST volume for this analysis; a bare `Context`
-    // render measured ~14 GB. 2 GB sits well above the former and far below the latter.
-    final long maxFormattedChars = 2_000_000_000L;
-
-    Logger pkg = Logger.getLogger("com.ibm.wala.cast.python");
-    DiscardingFormattingHandler handler = new DiscardingFormattingHandler();
-    handler.setLevel(Level.ALL);
-    Level oldLevel = pkg.getLevel();
-    boolean oldUseParentHandlers = pkg.getUseParentHandlers();
-    pkg.addHandler(handler);
-    pkg.setLevel(Level.FINEST);
-    // Count the FINEST volume here; don't also propagate it to the console handlers.
-    pkg.setUseParentHandlers(false);
-    try {
-      DiscardingFormattingHandler.reset();
-      testNlpgnnFullGeneration();
-      long volume = DiscardingFormattingHandler.totalChars();
-      assertTrue(
-          "Diagnostic FINEST volume "
-              + volume
-              + " chars exceeds the "
-              + maxFormattedChars
-              + "-char bound; a pointer-analysis or call-graph value is likely logged without"
-              + " Loggables.describe(...). See CONTRIBUTING.md's Diagnostic Logging section.",
-          volume < maxFormattedChars);
-    } finally {
-      pkg.removeHandler(handler);
-      pkg.setLevel(oldLevel);
-      pkg.setUseParentHandlers(oldUseParentHandlers);
-    }
   }
 
   /**
